+++ hexagonal/src/main/java/com/example/hexagonal/HexagonalApplication.java
@@ -6,8 +6,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class HexagonalApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(HexagonalApplication.class, args);
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(HexagonalApplication.class, args);
+    }
 
 }
