+++ hexagonal/src/main/java/com/example/hexagonal/domain/Customer.java
@@ -6,6 +6,7 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+
 /**
  * A user that interacts with the content
  * of movie hub
@@ -13,35 +14,35 @@ import javax.persistence.Id;
 @Entity
 public class Customer {
 
-	@Id
-	@GeneratedValue(strategy=GenerationType.AUTO)
-	private long userid;
-	private String name;
-	private String location;
-	private Date customerSince;
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private long userid;
+    private String name;
+    private String location;
+    private Date customerSince;
 
-	protected Customer() {
-	}
+    protected Customer() {
+    }
 
-	public Customer(String name, String location) {
-		this.name = name;
-		this.location = location;
-		this.customerSince = new Date();
-	}
+    public Customer(String name, String location) {
+        this.name = name;
+        this.location = location;
+        this.customerSince = new Date();
+    }
 
-	public long getUserid() {
-		return userid;
-	}
+    public long getUserid() {
+        return userid;
+    }
 
-	public String getName() {
-		return name;
-	}
+    public String getName() {
+        return name;
+    }
 
-	public String getLocation() {
-		return location;
-	}
+    public String getLocation() {
+        return location;
+    }
 
-	public Date getCustomerSince() {
-		return customerSince;
-	}
+    public Date getCustomerSince() {
+        return customerSince;
+    }
 }
