--- hexagonal/src/main/java/com/example/hexagonal/domain/CustomerRepository.java
@@ -1,7 +1,0 @@
-package com.example.hexagonal.domain;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface CustomerRepository extends CrudRepository<Customer, Long>{	
-	Customer findCustomerByUserid(long id);
-}
