--- hexagonal/src/main/java/com/example/hexagonal/domain/MovieRepository.java
@@ -1,7 +1,0 @@
-package com.example.hexagonal.domain;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface MovieRepository extends CrudRepository<Movie, Long>{
-	Movie findMovieByTitle(String name);	
-}
