+++ hexagonal/src/main/java/com/example/hexagonal/port/MovieHubService.java
@@ -1,0 +1,16 @@
+package com.example.hexagonal.port;
+
+import com.example.hexagonal.domain.Movie;
+
+/**
+ * exposes a point for interaction with the outside world
+ * and sets the protocol for adapter
+ *
+ */
+public interface MovieHubService {
+    Movie getMovieDetails(String movieTitle);
+
+    String registerNewUser(String userName, String userLocation);
+
+    String addMovie(int id, String title, int releaseYear, double imdbRating, int rottenTomatoesScore);
+}
