+++ hexagonal/src/main/java/com/example/hexagonal/repository/CustomerRepository.java
@@ -1,0 +1,8 @@
+package com.example.hexagonal.repository;
+
+import org.springframework.data.repository.CrudRepository;
+import com.example.hexagonal.domain.Customer;
+
+public interface CustomerRepository extends CrudRepository<Customer, Long> {
+    Customer findCustomerByUserid(long id);
+}
